+++ samples/demo/demo.js
@@ -4,7 +4,7 @@ var $ = require('jquery');
 var _ = require('underscore');
 var Backbone = require('backbone');
 Backbone.$ = $;
-var VDOMItemView = require('../index.js').VDOMItemView;
+var VDOMItemView = require('../../index.js').ItemView;
 
 var View = VDOMItemView.extend({
   template: _.template('<p>VIEW <%= cid %> </br> <b><%= text %></b></br><input class="input" value="<%= text %>"/></br><textarea class="input"><%= text %></textarea></p>'),
