+++ samples/perf/perf.html
