+++ samples/perf/perf.js
@@ -5,7 +5,7 @@ var _ = require('underscore');
 var Backbone = require('backbone');
 Backbone.$ = $;
 var Marionette = require('backbone.marionette');
-var VDOMItemView = require('../index.js').VDOMItemView;
+var VDOMItemView = require('../../index.js').ItemView;
 var React = require('react');
 
 // Change N to change the number of drawn circles.
